add another game: Tiger Stick and Chicken
how to play: enter either t, c, s, or b which stands for tiger, chicken, stick and bug. the tiger eats the chicken, the chicken eats the bug, the bug eats the stick and the stick hits the tiger. you will play against a computer but it isn't impossible to win.